feat: graduate discoveries page prototype
