+++ apps/web/src/features/discoveries/types.ts
@@ -1,0 +1,11 @@
+export type WeeklyActivityPoint = {
+  name: string;
+  value: number;
+};
+
+export type ArtistRelease = {
+  album: string;
+  releaseDate: string;
+  periodLabel: string;
+  image: string | null;
+};
