+++ apps/web/src/lib/query.ts
@@ -1,5 +1,6 @@
 import { albumsQueries } from "@/features/albums/queries";
 import { artistsQueries } from "@/features/artists/queries";
+import { discoveriesQueries } from "@/features/discoveries/queries";
 import { interactionsQueries } from "@/features/interactions/queries";
 import { listeningQueries } from "@/features/listening/queries";
 import { packagesQueries } from "@/features/packages/queries";
@@ -8,6 +9,7 @@ import { tracksQueries } from "@/features/tracks/queries";
 export const query = {
   albums: albumsQueries,
   artists: artistsQueries,
+  discoveries: discoveriesQueries,
   interactions: interactionsQueries,
   listeningHabits: listeningQueries,
   packages: packagesQueries,
